third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/108-expected.txt

This is a testharness.js-based test.
FAIL  scheduler: javascript URL in iframe assert_array_equals: property 1, expected "inline script #2" but got "iframe script #1"
Harness: the test ran to completion.
